fix: SharedPostsMenu 애니메이션 효과 수정 (#24)
