--- src/components/main-page/PostCard.tsx
@@ -1,117 +1,0 @@
-'use client';
-
-import styled from 'styled-components';
-
-import { HorizontalDivider } from '..';
-
-import { type SharedPost } from '@/entities/shared-post';
-
-const styles = {
-  container: styled.div`
-    width: 100%;
-
-    display: flex;
-    flex-direction: column;
-  `,
-  wrapper: styled.div`
-    width: 100%;
-    height: fit-content;
-
-    display: flex;
-    margin-bottom: 38px;
-  `,
-  thumbnail: styled.div<{ $url?: string }>`
-    width: 258px;
-    height: 200px;
-    flex-shrink: 0;
-    border-radius: 16px;
-    background: #f7f6f9;
-    background-image: ${({ $url }) =>
-      $url !== undefined ? `url("${$url}")` : 'none'};
-  `,
-  information: styled.div`
-    display: flex;
-    flex-grow: 1;
-    flex-direction: column;
-    gap: 21px;
-    margin-top: 14px;
-    margin-left: 22px;
-
-    h1 {
-      color: #000;
-      font-family: Inter;
-      font-size: 24px;
-      font-style: normal;
-      font-weight: 500;
-      line-height: normal;
-    }
-
-    p {
-      color: #000;
-      font-family: Inter;
-      font-size: 20px;
-      font-style: normal;
-      font-weight: 500;
-      line-height: normal;
-    }
-  `,
-  writerInformation: styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 16px;
-
-    img {
-      display: flex;
-      width: 100px;
-      height: 100px;
-      justify-content: center;
-      align-items: center;
-      flex-shrink: 0;
-
-      border-radius: 100px;
-      border: 1px solid #dcddea;
-      background: #c4c4c4;
-    }
-
-    p {
-      color: #000;
-      font-family: 'Noto Sans KR';
-      font-size: 24px;
-      font-style: normal;
-      font-weight: 500;
-      line-height: normal;
-    }
-  `,
-};
-
-function Thumbnail({ url }: { url?: string }) {
-  return <styles.thumbnail $url={url} />;
-}
-
-interface Prop {
-  url?: string;
-  post: SharedPost;
-  writer: {
-    name: string;
-  };
-}
-
-export function PostCard({ url, post, writer }: Prop) {
-  return (
-    <styles.container>
-      <styles.wrapper>
-        <Thumbnail url={url} />
-        <styles.information>
-          <h1>{post.title}</h1>
-          <p>{post.content}</p>
-        </styles.information>
-        <styles.writerInformation>
-          <img alt="" />
-          <p>{writer.name}</p>
-        </styles.writerInformation>
-      </styles.wrapper>
-      <HorizontalDivider />
-    </styles.container>
-  );
-}
