+++ src/components/main-page/index.ts
@@ -1,2 +1,1 @@
 export * from './UserCard';
-export * from './PostCard';
