+++ src/components/shared-posts/PostCard.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import { type SharedPost } from '@/entities/shared-post';
+
+const styles = {};
+
+interface Prop {
+  url?: string;
+  post: SharedPost;
+  writer: {
+    name: string;
+  };
+}
+
+export function PostCard() {
+  return <></>;
+}
