+++ src/components/shared-posts/index.ts
@@ -1,2 +1,3 @@
 export * from './SharedPostsMenu';
 export * from './SharedPostsFilter';
+export * from './PostCard';
